reorganisation and improvment of react folder structure
