+++ reactmock/src/App.js
@@ -1,12 +1,12 @@
 import { useState } from "react";
-import SearchBar from "./components/SearchBar";
-import searchContent from "./api";
-import ResultList from "./components/ResultList";
+import SearchBar from "./components/results/SearchBar";
+import SearchContent from "data/api";
+import ResultList from "./components/results/ResultList";
 
 function App() {
   const [remedies, setRemedies] = useState([]);
   const handleSubmit = async (term) => {
-    const result = await searchContent(term);
+    const result = await SearchContent(term);
 
     setRemedies(result);
   };
