+++ reactmock/src/components/results/SearchBar.js
